+++ src/data/Thread.js
@@ -1,7 +1,12 @@
+const moment = require('moment');
+
 const bot = require('../bot');
 const knex = require('../knex');
 const utils = require('../utils');
+const config = require('../config');
 const attachments = require('./attachments');
+
+const ThreadMessage = require('./ThreadMessage');
 
 const {THREAD_MESSAGE_TYPE, THREAD_STATUS} = require('./constants');
 
@@ -12,17 +17,16 @@ const {THREAD_MESSAGE_TYPE, THREAD_STATUS} = require('./constants');
  * @property {String} user_name
  * @property {String} channel_id
  * @property {String} created_at
- * @property {Boolean} _wasCreated
  */
 class Thread {
   constructor(props) {
-    Object.assign(this, {_wasCreated: false}, props);
+    Object.assign(this, props);
   }
 
   /**
-   * @param {Eris.Member} moderator
+   * @param {Eris~Member} moderator
    * @param {String} text
-   * @param {Eris.Attachment[]} replyAttachments
+   * @param {Eris~Attachment[]} replyAttachments
    * @param {Boolean} isAnonymous
    * @returns {Promise<void>}
    */
@@ -81,6 +85,7 @@ class Thread {
       user_id: moderator.id,
       user_name: logModUsername,
       body: logContent,
+      is_anonymous: (isAnonymous ? 1 : 0),
       original_message_id: originalMessage.id
     });
   }
@@ -92,7 +97,12 @@ class Thread {
   async receiveUserReply(msg) {
     const timestamp = utils.getTimestamp();
 
-    let threadContent = `[${timestamp}] « **${msg.author.username}#${msg.author.discriminator}:** ${msg.content}`;
+    let content = msg.content;
+    if (msg.content.trim() === '' && msg.embeds.length) {
+      content = '<message contains embeds>';
+    }
+
+    let threadContent = `[${timestamp}] « **${msg.author.username}#${msg.author.discriminator}:** ${content}`;
     let logContent = msg.content;
     let finalThreadContent;
     let attachmentSavePromise;
@@ -113,6 +123,7 @@ class Thread {
       user_id: this.user_id,
       user_name: `${msg.author.username}#${msg.author.discriminator}`,
       body: logContent,
+      is_anonymous: 0,
       original_message_id: msg.id
     });
 
@@ -128,8 +139,7 @@ class Thread {
    * @returns {Promise<Eris.Message>}
    */
   async postToThreadChannel(text, file = null) {
-    const channel = bot.getChannel(this.channel_id);
-    return channel.createMessage(text, file);
+    return bot.createMessage(this.channel_id, text, file);
   }
 
   /**
@@ -143,8 +153,56 @@ class Thread {
       user_id: null,
       user_name: '',
       body: text,
+      is_anonymous: 0,
       original_message_id: msg.id
     });
+  }
+
+  /**
+   * @param {String} text
+   * @returns {Promise<void>}
+   */
+  async postNonLogMessage(text) {
+    await this.postToThreadChannel(text);
+  }
+
+  /**
+   * @param {Eris.Message} msg
+   * @returns {Promise<void>}
+   */
+  async saveChatMessage(msg) {
+    return this.addThreadMessageToDB({
+      message_type: THREAD_MESSAGE_TYPE.CHAT,
+      user_id: msg.author.id,
+      user_name: `${msg.author.username}#${msg.author.discriminator}`,
+      body: msg.content,
+      is_anonymous: 0,
+      original_message_id: msg.id
+    });
+  }
+
+  /**
+   * @param {Eris.Message} msg
+   * @returns {Promise<void>}
+   */
+  async updateChatMessage(msg) {
+    await knex('thread_messages')
+      .where('thread_id', this.id)
+      .where('original_message_id', msg.id)
+      .update({
+        content: msg.content
+      });
+  }
+
+  /**
+   * @param {String} messageId
+   * @returns {Promise<void>}
+   */
+  async deleteChatMessage(messageId) {
+    await knex('thread_messages')
+      .where('thread_id', this.id)
+      .where('original_message_id', messageId)
+      .delete();
   }
 
   /**
@@ -155,14 +213,29 @@ class Thread {
     await knex('thread_messages').insert({
       thread_id: this.id,
       created_at: moment.utc().format('YYYY-MM-DD HH:mm:ss'),
+      is_anonymous: 0,
       ...data
     });
+  }
+
+  /**
+   * @returns {Promise<ThreadMessage[]>}
+   */
+  async getThreadMessages() {
+    const threadMessages = await knex('thread_messages')
+      .where('thread_id', this.id)
+      .orderBy('created_at', 'ASC')
+      .orderBy('id', 'ASC')
+      .select();
+
+    return threadMessages.map(row => new ThreadMessage(row));
   }
 
   /**
    * @returns {Promise<void>}
    */
   async close() {
+    console.log(`Closing thread ${this.id}`);
     await this.postToThreadChannel('Closing thread...');
 
     // Update DB status
@@ -173,9 +246,10 @@ class Thread {
       });
 
     // Delete channel
+    console.log(`Deleting channel ${this.channel_id}`);
     const channel = bot.getChannel(this.channel_id);
     if (channel) {
-      channel.delete('Thread closed');
+      await channel.delete('Thread closed');
     }
   }