+++ src/main.js
@@ -5,48 +5,28 @@ const config = require('./config');
 const bot = require('./bot');
 const Queue = require('./queue');
 const utils = require('./utils');
+const threadUtils = require('./threadUtils');
 const blocked = require('./data/blocked');
 const threads = require('./data/threads');
-const snippets = require('./data/snippets');
-const webserver = require('./webserver');
-const greeting = require('./greeting');
-const Thread = require('./data/Thread');
+
+const snippets = require('./plugins/snippets');
+const webserver = require('./plugins/webserver');
+const greeting = require('./plugins/greeting');
 
 const messageQueue = new Queue();
 
-/**
- * @callback CommandHandlerCB
- * @interface
- * @param {Eris~Message} msg
- * @param {Array} args
- * @param {Thread} thread
- * @return void
- */
-
-/**
- * Adds a command that can only be triggered on the inbox server.
- * Command handlers added with this function also get the thread the message was posted in as a third argument, if any.
- * @param {String} cmd
- * @param {CommandHandlerCB} commandHandler
- * @param {Eris~CommandOptions} opts
- */
-function addInboxServerCommand(cmd, commandHandler, opts) {
-  bot.registerCommand(cmd, async (msg, args) => {
-    if (! messageIsOnInboxServer(msg)) return;
-    if (! isStaff(msg.member)) return;
-
-    const thread = await threads.findByChannelId(msg.channel.id);
-    commandHandler(msg, args, thread);
-  }, opts);
-}
+const addInboxServerCommand = (...args) => threadUtils.addInboxServerCommand(bot, ...args);
 
 // Once the bot has connected, set the status/"playing" message
 bot.on('ready', () => {
   bot.editStatus(null, {name: config.status});
-  console.log('Bot started, listening to DMs');
 });
 
-// Handle moderator messages in thread channels
+/**
+ * When a moderator posts in a modmail thread...
+ * 1) If alwaysReply is enabled, reply to the user
+ * 2) If alwaysReply is disabled, save that message as a chat message in the thread
+ */
 bot.on('messageCreate', async msg => {
   if (! utils.messageIsOnInboxServer(msg)) return;
   if (! utils.isStaff(msg)) return;
@@ -62,17 +42,81 @@ bot.on('messageCreate', async msg => {
     msg.delete();
   } else {
     // Otherwise just save the messages as "chat" in the logs
-    thread.addThreadMessageToDB({
-      message_type: threads.THREAD_MESSAGE_TYPE.CHAT,
-      user_id: msg.author.id,
-      user_name: `${msg.author.username}#${msg.author.discriminator}`,
-      body: msg.content,
-      original_message_id: msg.id
-    });
+    thread.saveChatMessage(msg);
   }
 });
 
-// If the bot is mentioned on the main server, post a log message about it
+/**
+ * When we get a private message...
+ * 1) Find the open modmail thread for this user, or create a new one
+ * 2) Post the message as a user reply in the thread
+ */
+bot.on('messageCreate', async msg => {
+  if (! (msg.channel instanceof Eris.PrivateChannel)) return;
+  if (msg.author.bot) return;
+  if (msg.type !== 0) return; // Ignore pins etc.
+
+  if (await blocked.isBlocked(msg.author.id)) return;
+
+  // Private message handling is queued so e.g. multiple message in quick succession don't result in multiple channels being created
+  messageQueue.add(async () => {
+    const thread = await threads.findOrCreateThreadForUser(msg.author);
+    await thread.receiveUserReply(msg);
+  });
+});
+
+/**
+ * When a message is edited...
+ * 1) If that message was in DMs, and we have a thread open with that user, post the edit as a system message in the thread
+ * 2) If that message was moderator chatter in the thread, update the corresponding chat message in the DB
+ */
+bot.on('messageUpdate', async (msg, oldMessage) => {
+  if (msg.author.bot) return;
+  if (await blocked.isBlocked(msg.author.id)) return;
+
+  let oldContent = oldMessage.content;
+  const newContent = msg.content;
+
+  // Old message content doesn't persist between bot restarts
+  if (oldContent == null) oldContent = '*Unavailable due to bot restart*';
+
+  // Ignore bogus edit events with no changes
+  if (newContent.trim() === oldContent.trim()) return;
+
+  // 1) Edit in DMs
+  if (msg.channel instanceof Eris.PrivateChannel) {
+    const thread = await threads.findOpenThreadByUserId(msg.author.id);
+    const editMessage = utils.disableLinkPreviews(`**The user edited their message:**\n\`B:\` ${oldContent}\n\`A:\` ${newContent}`);
+
+    thread.postSystemMessage(editMessage);
+  }
+
+  // 2) Edit in the thread
+  else if (utils.messageIsOnInboxServer(msg) && utils.isStaff(msg.member)) {
+    const thread = await threads.findOpenThreadByChannelId(msg.channel.id);
+    if (! thread) return;
+
+    thread.updateChatMessage(msg);
+  }
+});
+
+/**
+ * When a staff message is deleted in a modmail thread, delete it from the database as well
+ */
+bot.on('messageDelete', async msg => {
+  if (msg.author.bot) return;
+  if (! utils.messageIsOnInboxServer(msg)) return;
+  if (! utils.isStaff(msg.member)) return;
+
+  const thread = await threads.findOpenThreadByChannelId(msg.channel.id);
+  if (! thread) return;
+
+  thread.deleteChatMessage(msg.id);
+});
+
+/**
+ * When the bot is mentioned on the main server, ping staff in the log channel about it
+ */
 bot.on('messageCreate', async msg => {
   if (! utils.messageIsOnMainServer(msg)) return;
   if (! msg.mentions.some(user => user.id === bot.user.id)) return;
@@ -89,62 +133,25 @@ bot.on('messageCreate', async msg => {
   });
 });
 
-// When we get a private message, forward the contents to the corresponding modmail thread
-bot.on('messageCreate', async msg => {
-  if (! (msg.channel instanceof Eris.PrivateChannel)) return;
-  if (msg.author.id === bot.user.id) return;
-
-  if (await blocked.isBlocked(msg.author.id)) return;
-
-  // Private message handling is queued so e.g. multiple message in quick succession don't result in multiple channels being created
-  messageQueue.add(async () => {
-    let thread = await threads.findOpenThreadByUserId(msg.author.id);
-    if (! thread) {
-      thread = await threads.createNewThreadForUser(msg.author, msg);
-    }
-
-    thread.receiveUserReply(msg);
-  });
-});
-
-// Edits in DMs
-bot.on('messageUpdate', async (msg, oldMessage) => {
-  if (! (msg.channel instanceof Eris.PrivateChannel)) return;
-  if (msg.author.id === bot.user.id) return;
-
-  if (await blocked.isBlocked(msg.author.id)) return;
-
-  let oldContent = oldMessage.content;
-  const newContent = msg.content;
-
-  // Old message content doesn't persist between bot restarts
-  if (oldContent == null) oldContent = '*Unavailable due to bot restart*';
-
-  // Ignore bogus edit events with no changes
-  if (newContent.trim() === oldContent.trim()) return;
-
-  const thread = await threads.createNewThreadForUser(msg.author);
-  if (! thread) return;
-
-  const editMessage = utils.disableLinkPreviews(`**The user edited their message:**\n\`B:\` ${oldContent}\n\`A:\` ${newContent}`);
-  bot.createMessage(thread.channelId, editMessage);
-});
-
 // Mods can reply to modmail threads using !r or !reply
 // These messages get relayed back to the DM thread between the bot and the user
-addInboxServerCommand('reply', (msg, args, thread) => {
+addInboxServerCommand('reply', async (msg, args, thread) => {
   if (! thread) return;
+
   const text = args.join(' ').trim();
-  thread.replyToUser(msg.member, text, msg.attachments, false);
+  await thread.replyToUser(msg.member, text, msg.attachments, false);
+  msg.delete();
 });
 
 bot.registerCommandAlias('r', 'reply');
 
 // Anonymous replies only show the role, not the username
-addInboxServerCommand('anonreply', (msg, args, thread) => {
+addInboxServerCommand('anonreply', async (msg, args, thread) => {
   if (! thread) return;
+
   const text = args.join(' ').trim();
-  thread.replyToUser(msg.member, text, msg.attachments, true);
+  await thread.replyToUser(msg.member, text, msg.attachments, true);
+  msg.delete();
 });
 
 bot.registerCommandAlias('ar', 'anonreply');
@@ -168,7 +175,7 @@ addInboxServerCommand('block', (msg, args, thread) => {
     block(userId);
   } else if (thread) {
     // Calling !block without args in a modmail thread blocks the user of that thread
-    block(thread.userId);
+    block(thread.user_id);
   }
 });
 
@@ -185,7 +192,7 @@ addInboxServerCommand('unblock', (msg, args, thread) => {
     unblock(userId);
   } else if (thread) {
     // Calling !unblock without args in a modmail thread unblocks the user of that thread
-    unblock(thread.userId);
+    unblock(thread.user_id);
   }
 });
 
@@ -217,105 +224,21 @@ addInboxServerCommand('logs', (msg, args, thread) => {
     getLogs(userId);
   } else if (thread) {
     // Calling !logs without args in a modmail thread returns the logs of the user of that thread
-    getLogs(thread.userId);
+    getLogs(thread.user_id);
   }
-});
-
-// Snippets
-bot.on('messageCreate', async msg => {
-  if (! utils.messageIsOnInboxServer(msg)) return;
-  if (! utils.isStaff(msg.member)) return;
-
-  if (msg.author.bot) return;
-  if (! msg.content) return;
-  if (! msg.content.startsWith(config.snippetPrefix)) return;
-
-  const shortcut = msg.content.replace(config.snippetPrefix, '').toLowerCase();
-  const snippet = await snippets.get(shortcut);
-  if (! snippet) return;
-
-  reply(msg, snippet.text, snippet.isAnonymous);
-});
-
-// Show or add a snippet
-addInboxServerCommand('snippet', async (msg, args) => {
-  const shortcut = args[0];
-  if (! shortcut) return
-
-  const text = args.slice(1).join(' ').trim();
-  const snippet = await snippets.get(shortcut);
-
-  if (snippet) {
-    if (text) {
-      // If the snippet exists and we're trying to create a new one, inform the user the snippet already exists
-      msg.channel.createMessage(`Snippet "${shortcut}" already exists! You can edit or delete it with ${prefix}edit_snippet and ${prefix}delete_snippet respectively.`);
-    } else {
-      // If the snippet exists and we're NOT trying to create a new one, show info about the existing snippet
-      msg.channel.createMessage(`\`${config.snippetPrefix}${shortcut}\` replies ${snippet.isAnonymous ? 'anonymously ' : ''}with:\n${snippet.text}`);
-    }
-  } else {
-    if (text) {
-      // If the snippet doesn't exist and the user wants to create it, create it
-      await snippets.add(shortcut, text, false);
-      msg.channel.createMessage(`Snippet "${shortcut}" created!`);
-    } else {
-      // If the snippet doesn't exist and the user isn't trying to create it, inform them how to create it
-      msg.channel.createMessage(`Snippet "${shortcut}" doesn't exist! You can create it with \`${prefix}snippet ${shortcut} text\``);
-    }
-  }
-});
-
-bot.registerCommandAlias('s', 'snippet');
-
-addInboxServerCommand('delete_snippet', async (msg, args) => {
-  const shortcut = args[0];
-  if (! shortcut) return;
-
-  const snippet = await snippets.get(shortcut);
-  if (! snippet) {
-    msg.channel.createMessage(`Snippet "${shortcut}" doesn't exist!`);
-    return;
-  }
-
-  await snippets.del(shortcut);
-  msg.channel.createMessage(`Snippet "${shortcut}" deleted!`);
-});
-
-bot.registerCommandAlias('ds', 'delete_snippet');
-
-addInboxServerCommand('edit_snippet', async (msg, args) => {
-  const shortcut = args[0];
-  if (! shortcut) return;
-
-  const text = args.slice(1).join(' ').trim();
-  if (! text) return;
-
-  const snippet = await snippets.get(shortcut);
-  if (! snippet) {
-    msg.channel.createMessage(`Snippet "${shortcut}" doesn't exist!`);
-    return;
-  }
-
-  await snippets.del(shortcut);
-  await snippets.add(shortcut, text, snippet.isAnonymous);
-
-  msg.channel.createMessage(`Snippet "${shortcut}" edited!`);
-});
-
-bot.registerCommandAlias('es', 'edit_snippet');
-
-addInboxServerCommand('snippets', async msg => {
-  const allSnippets = await snippets.all();
-  const shortcuts = Object.keys(allSnippets);
-  shortcuts.sort();
-
-  msg.channel.createMessage(`Available snippets (prefix ${config.snippetPrefix}):\n${shortcuts.join(', ')}`);
 });
 
 module.exports = {
-  start() {
-    bot.connect();
-    // webserver.run();
-    greeting.init(bot);
+  async start() {
+    // Load plugins
+    console.log('Loading plugins...');
+    await snippets(bot);
+    await greeting(bot);
+    await webserver(bot);
+
+    console.log('Connecting to Discord...');
+    await bot.connect();
+
+    console.log('Done! Now listening to DMs.');
   }
 };