+++ src/utils.js
@@ -13,12 +13,18 @@ let inboxGuild = null;
 let mainGuild = null;
 let logChannel = null;
 
+/**
+ * @returns {Eris~Guild}
+ */
 function getInboxGuild() {
   if (! inboxGuild) inboxGuild = bot.guilds.find(g => g.id === config.mailGuildId);
   if (! inboxGuild) throw new BotError('The bot is not on the modmail (inbox) server!');
   return inboxGuild;
 }
 
+/**
+ * @returns {Eris~Guild}
+ */
 function getMainGuild() {
   if (! mainGuild) mainGuild = bot.guilds.find(g => g.id === config.mainGuildId);
   if (! mainGuild) console.warn('[WARN] The bot is not on the main server! If this is intentional, you can ignore this warning.');
@@ -28,7 +34,7 @@ function getMainGuild() {
 /**
  * Returns the designated log channel, or the default channel if none is set
  * @param bot
- * @returns {object}
+ * @returns {Eris~TextChannel}
  */
 function getLogChannel() {
   const inboxGuild = getInboxGuild();
@@ -151,8 +157,8 @@ async function getSelfUrl(path = '') {
 
 /**
  * Returns the highest hoisted role of the given member
- * @param {Eris.Member} member
- * @returns {Eris.Role}
+ * @param {Eris~Member} member
+ * @returns {Eris~Role}
  */
 function getMainRole(member) {
   const roles = member.roles.map(id => member.guild.roles.get(id));