+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v2.0.0
+* Rewrote large parts of the code to be more modular and maintainable. There may be some new bugs because of this - please report them through GitHub issues if you encounter any!
+* Threads, logs, and snippets are now stored in an SQLite database. The bot will migrate old data on the first run.
+* Fixed system messages like pins in DMs being relayed to the thread
+* Fixed channels sometimes being created without a category
+
 ## Sep 22, 2017
 * Added `newThreadCategoryId` option. This option can be set to a category ID to place all new threads in that category.